src: add support to multi testing prompts
We can simply add more input lines and then put the prompts in the
file as different lines. But this solution does not support arbitrary
number of test prompts as each now group of prompts need a new input
line.

So we make the decision to support sub format to split prompt groups
in the same input prompts, which is more flexible.