fix: reset useless config value
The face_crop_aug_range means the dataset loader crop the input images
according to the detected face position, so that the LoRA can focus more
on the faces of the image.

However, there is no integrated face detector in the repo and our target
is not a specific face reconstruction task. So, we remove reset the conf
to None.